refactor: delete window abstraction
